+++ browser/data-browser/src/components/Thumbnail.tsx
@@ -1,32 +1,32 @@
 import { styled } from 'styled-components';
+import { Image } from '@tomic/react';
 import { InnerWrapper } from '../views/FolderPage/GridItem/components';
 
 interface ThumbnailProps {
-  src: string | undefined;
+  subject: string;
   style?: React.CSSProperties | undefined;
 }
 
-export function Thumbnail({ src, style }: ThumbnailProps): JSX.Element {
-  if (src === undefined) {
-    return <TextWrapper>No preview available</TextWrapper>;
-  }
-
+export function Thumbnail({ subject, style }: ThumbnailProps): JSX.Element {
   return (
     <InnerWrapper>
-      <Image src={src} alt='' loading='lazy' style={style} />
+      <StyledImage
+        subject={subject}
+        alt=''
+        loading='lazy'
+        style={style}
+        sizeIndication={{
+          '55px': 100,
+          default: 30,
+        }}
+      />
     </InnerWrapper>
   );
 }
 
-const Image = styled.img`
+const StyledImage = styled(Image)`
   width: 100%;
   height: 100%;
   object-fit: cover;
   object-position: center;
-`;
-
-const TextWrapper = styled(InnerWrapper)`
-  display: grid;
-  place-items: center;
-  color: ${p => p.theme.colors.textLight};
 `;