+++ browser/data-browser/src/views/TablePage/EditorCells/ResourceCells/FileCell.tsx
@@ -1,4 +1,4 @@
-import { properties, useResource, useString, useTitle } from '@tomic/react';
+import { useResource, useString, useTitle, Image, server } from '@tomic/react';
 
 import { styled } from 'styled-components';
 import { getFileIcon, imageMimeTypes } from '../../../../helpers/filetypes';
@@ -8,8 +8,7 @@ import { SimpleResourceLink } from './SimpleResourceLink';
 export function FileCell({ subject }: ResourceCellProps) {
   const resource = useResource(subject);
   const [title] = useTitle(resource);
-  const [mimeType] = useString(resource, properties.file.mimetype);
-  const [downloadUrl] = useString(resource, properties.file.downloadUrl);
+  const [mimeType] = useString(resource, server.properties.mimetype);
 
   const isImage = imageMimeTypes.has(mimeType ?? '');
   const Icon = getFileIcon(mimeType ?? '');
@@ -18,7 +17,12 @@ export function FileCell({ subject }: ResourceCellProps) {
     <Wrapper>
       {isImage ? (
         <StyledLink resource={resource} tabIndex={-1} aria-hidden>
-          <Img src={downloadUrl} alt={title} loading='lazy' />
+          <Img
+            subject={subject}
+            alt={title}
+            loading='lazy'
+            sizeIndication={'100px'}
+          />
         </StyledLink>
       ) : (
         <Icon />
@@ -28,11 +32,13 @@ export function FileCell({ subject }: ResourceCellProps) {
   );
 }
 
-const Img = styled.img`
+const Img = styled(Image)`
   width: calc(var(--table-row-height) - 6px);
+  height: calc(var(--table-row-height) - 6px);
   aspect-ratio: 1/1;
   object-fit: cover;
   border-radius: 5px;
+  vertical-align: middle;
 `;
 
 const Wrapper = styled.div`
@@ -45,4 +51,5 @@ const Wrapper = styled.div`
 const StyledLink = styled(SimpleResourceLink)`
   display: flex;
   height: fit-content;
+  align-items: center;
 `;