+++ browser/svelte/src/lib/components/Image/index.ts
@@ -1,0 +1,2 @@
+export { default as Image } from './Image.svelte';
+export { type SizeIndication } from './imageHelpers.js';