+++ browser/svelte/src/routes/+page.svelte
@@ -1,0 +1,8 @@
+<h1>@tomic/svelte exaples</h1>
+
+<h2>Components</h2>
+<ul>
+	<li>
+		<a href="/examples/image">{'<'}Image /{'>'}</a>
+	</li>
+</ul>