+++ browser/svelte/vite.config.ts
@@ -1,43 +1,12 @@
-import { defineConfig } from 'vite';
-
-import typescript from '@rollup/plugin-typescript';
-import path from 'path';
-import { typescriptPaths } from 'rollup-plugin-typescript-paths';
+import { sveltekit } from '@sveltejs/kit/vite';
+import { defineConfig } from 'vitest/config';
 
 export default defineConfig({
-  plugins: [],
-  resolve: {
-    alias: [
-      {
-        find: '~',
-        replacement: path.resolve(__dirname, './src'),
-      },
-    ],
-  },
-  server: {
-    port: 5175,
-  },
-  build: {
-    manifest: true,
-    minify: true,
-    reportCompressedSize: true,
-    lib: {
-      entry: path.resolve(__dirname, 'src/index.ts'),
-      fileName: 'index',
-      formats: ['es', 'cjs'],
-    },
-    rollupOptions: {
-      external: [],
-      plugins: [
-        typescriptPaths({
-          preserveExtensions: true,
-        }),
-        typescript({
-          sourceMap: false,
-          declaration: true,
-          outDir: 'dist',
-        }),
-      ],
-    },
-  },
+	plugins: [sveltekit()],
+	test: {
+		include: ['src/**/*.{test,spec}.{js,ts}']
+	},
+	server: {
+		port: 5175
+	}
 });